refactor(chat): split chat_runtime.py into focused modules (#76)
Split chat_runtime.py (372 lines) into:
- model_resolution.py (102 lines): model string parsing, env resolution, provider detection
- toolset_builder.py (196 lines): toolset construction, strict tool wrapping
- chat_runtime.py (94 lines): agent construction + runtime singleton

All imports updated across codebase and tests.
ci.yml source_to_spec mappings added for new files.
