+++ tests/test_chat_runtime.py
@@ -1,4 +1,4 @@
-"""Tests for chat_runtime: build_agent, build_toolsets, resolve_workspace_root, required_env."""
+"""Tests for runtime helpers split across chat modules."""
 
 from __future__ import annotations
 
@@ -36,13 +36,13 @@ class TestRequiredEnv:
     """Tests for required_env raising on missing vars."""
 
     def test_returns_value_when_set(self) -> None:
-        from chat_runtime import required_env
+        from model_resolution import required_env
 
         with patch.dict(os.environ, {"TEST_VAR_XYZ": "hello"}):
             assert required_env("TEST_VAR_XYZ") == "hello"
 
     def test_raises_system_exit_when_missing(self) -> None:
-        from chat_runtime import required_env
+        from model_resolution import required_env
 
         with patch.dict(os.environ, {}, clear=False):
             os.environ.pop("MISSING_VAR_ABC", None)
@@ -50,7 +50,7 @@ class TestRequiredEnv:
                 required_env("MISSING_VAR_ABC")
 
     def test_raises_on_empty_string(self) -> None:
-        from chat_runtime import required_env
+        from model_resolution import required_env
 
         with (
             patch.dict(os.environ, {"EMPTY_VAR": ""}),
@@ -63,7 +63,7 @@ class TestResolveWorkspaceRoot:
     """Tests for resolve_workspace_root with various env configs."""
 
     def test_absolute_path_used_directly(self, tmp_path: Path) -> None:
-        from chat_runtime import resolve_workspace_root
+        from toolset_builder import resolve_workspace_root
 
         target = tmp_path / "workspace"
         with patch.dict(os.environ, {"AGENT_WORKSPACE_ROOT": str(target)}):
@@ -72,7 +72,7 @@ class TestResolveWorkspaceRoot:
             assert result.is_dir()
 
     def test_relative_path_resolved_from_module(self) -> None:
-        from chat_runtime import resolve_workspace_root
+        from toolset_builder import resolve_workspace_root
 
         with patch.dict(os.environ, {"AGENT_WORKSPACE_ROOT": "data/test-ws"}):
             result = resolve_workspace_root()
@@ -81,7 +81,7 @@ class TestResolveWorkspaceRoot:
             assert result.is_dir()
 
     def test_default_when_unset(self) -> None:
-        from chat_runtime import resolve_workspace_root
+        from toolset_builder import resolve_workspace_root
 
         with patch.dict(os.environ, {}, clear=False):
             os.environ.pop("AGENT_WORKSPACE_ROOT", None)
@@ -90,7 +90,7 @@ class TestResolveWorkspaceRoot:
             assert result.name == "agent-workspace"
 
     def test_creates_directory(self, tmp_path: Path) -> None:
-        from chat_runtime import resolve_workspace_root
+        from toolset_builder import resolve_workspace_root
 
         target = tmp_path / "new" / "deep" / "ws"
         with patch.dict(os.environ, {"AGENT_WORKSPACE_ROOT": str(target)}):
@@ -107,7 +107,7 @@ class TestBuildToolsets:
         monkeypatch.setenv("ANTHROPIC_API_KEY", "test-key")
 
     def test_returns_tuple(self) -> None:
-        from chat_runtime import build_toolsets
+        from toolset_builder import build_toolsets
 
         result = build_toolsets()
         assert isinstance(result, tuple)
@@ -117,7 +117,7 @@ class TestBuildToolsets:
         assert len(system_prompt) > 0
 
     def test_toolset_count_without_memory(self) -> None:
-        from chat_runtime import build_toolsets
+        from toolset_builder import build_toolsets
 
         with patch.dict(os.environ, {"ENABLE_EXECUTE": ""}):
             toolsets, _ = build_toolsets()
@@ -126,7 +126,7 @@ class TestBuildToolsets:
             assert len(toolsets) >= min_toolsets
 
     def test_toolset_count_with_memory(self, tmp_path: Path) -> None:
-        from chat_runtime import build_toolsets
+        from toolset_builder import build_toolsets
 
         db_path = str(tmp_path / "mem.sqlite")
         from memory_store import init_memory_store
@@ -139,14 +139,14 @@ class TestBuildToolsets:
 
     def test_exec_instructions_absent_when_disabled(self) -> None:
         with patch.dict(os.environ, {"ENABLE_EXECUTE": ""}):
-            from chat_runtime import build_toolsets
+            from toolset_builder import build_toolsets
 
             _, prompt = build_toolsets()
             assert "## Shell execution" not in prompt
 
     def test_exec_instructions_present_when_enabled(self) -> None:
         with patch.dict(os.environ, {"ENABLE_EXECUTE": "true"}):
-            from chat_runtime import build_toolsets
+            from toolset_builder import build_toolsets
 
             _, prompt = build_toolsets()
             assert "## Shell execution" in prompt
