+++ tests/test_model_settings.py
@@ -7,7 +7,8 @@ from unittest.mock import patch
 
 from pydantic_ai.settings import ModelSettings
 
-from chat_runtime import AgentOptions, build_agent, build_model_settings
+from chat_runtime import AgentOptions, build_agent
+from model_resolution import build_model_settings
 
 _TEMP_HALF = 0.5
 _TEMP_LOW = 0.3
